+++ lib/src/generator/entity_generator.dart
@@ -12,6 +12,7 @@ class EntityGenerator {
       name: reqAttrValue(node, "name"),
       copyWith: asBool(optAttrValue(node, "copyWith", "true")),
       generateEntity: asBool(optAttrValue(node, "generateEntity", "true")),
+      apiProxies: asSplitList(attrValue(node, "apiProxy"))
     );
     final serializers = attrValue(node, "serializers");
     if (serializers != null && serializers.trim().isNotEmpty) {