+++ lib/src/generator/model_generator.dart
@@ -1,4 +1,6 @@
 import 'package:entity_serializer/src/common/xml_helper.dart';
+import 'package:entity_serializer/src/generator/api_proxy_generator.dart';
+import 'package:entity_serializer/src/model/api_proxy_info.dart';
 
 import 'entity_generator.dart';
 import 'import_generator.dart';
@@ -13,6 +15,7 @@ class Model {
   final List<Serializer> serializers = [];
   final List<Entity> entities = [];
   final List<Import> imports = [];
+  final List<ApiProxyInfo> proxies = [];
 
   bool get isNotEmpty => serializers.isNotEmpty || entities.isNotEmpty;
 }
@@ -38,8 +41,7 @@ class ModelGenerator {
     } else if (rootName == "entities") {
       _handleEntitiesRoot(from, model);
     } else {
-      throw Exception(
-          "Unsupported root: '${from.rootElement.name.toString()}'");
+      throw Exception("Unsupported root: '${from.rootElement.name.toString()}'");
     }
   }
 
@@ -63,8 +65,16 @@ class ModelGenerator {
     for (final node in from.rootElement.childElements) {
       final name = node.name.toString().toLowerCase();
       if (name == "serializer") {
-        model.serializers.add(SerializerGenerator.parseNode(node));
+        final ser = SerializerGenerator.parseNode(node);
+        if (model.serializers.any((element) => element.name == ser.name)) {
+          throw Exception("Serializer with name '${ser.name} is already defined!");
+        }
+        model.serializers.add(ser);
       } else if (name == "serializertemplate") {
+        final ser = SerializerGenerator.parseNode(node);
+        if (serializerTemplates.any((element) => element.name == ser.name)) {
+          throw Exception("Serializer template with name '${ser.name} is already defined!");
+        }
         serializerTemplates.add(SerializerGenerator.parseNode(node));
       } else {
         throw Exception("Unknown node with name: '${node.name.toString()}'");
@@ -80,9 +90,19 @@ class ModelGenerator {
     for (final node in from.rootElement.childElements) {
       final name = node.name.toString().toLowerCase();
       if (name == "class") {
-        model.entities.add(EntityGenerator.parseNode(node));
+        final cls = EntityGenerator.parseNode(node);
+        if (model.entities.any((element) => element.name == cls.name)) {
+          throw Exception("Entity with name '${cls.name} is already defined!");
+        }
+        model.entities.add(cls);
       } else if (name == "import") {
         model.imports.add(ImportGenerator.parseNode(node));
+      } else if (name == "proxy") {
+        final proxy = ApiProxyGenerator.parseNode(node);
+        if (model.proxies.any((element) => element.alias == proxy.alias)) {
+          throw Exception("Api proxy with name '${proxy.alias} is already defined!");
+        }
+        model.proxies.add(proxy);
       } else {
         throw Exception("Unknown node with name: '${node.name.toString()}'");
       }
@@ -94,13 +114,31 @@ class ModelGenerator {
     for (final node in from.rootElement.childElements) {
       final name = node.name.toString().toLowerCase();
       if (name == "class") {
-        model.entities.add(EntityGenerator.parseNode(node));
+        final cls = EntityGenerator.parseNode(node);
+        if (model.entities.any((element) => element.name == cls.name)) {
+          throw Exception("Entity with name '${cls.name} is already defined!");
+        }
+        model.entities.add(cls);
       } else if (name == "serializer") {
-        model.serializers.add(SerializerGenerator.parseNode(node));
+        final ser = SerializerGenerator.parseNode(node);
+        if (model.serializers.any((element) => element.name == ser.name)) {
+          throw Exception("Serializer with name '${ser.name} is already defined!");
+        }
+        model.serializers.add(ser);
       } else if (name == "serializertemplate") {
+        final ser = SerializerGenerator.parseNode(node);
+        if (serializerTemplates.any((element) => element.name == ser.name)) {
+          throw Exception("Serializer template with name '${ser.name} is already defined!");
+        }
         serializerTemplates.add(SerializerGenerator.parseNode(node));
       } else if (name == "import") {
         model.imports.add(ImportGenerator.parseNode(node));
+      } else if (name == "proxy") {
+        final proxy = ApiProxyGenerator.parseNode(node);
+        if (model.proxies.any((element) => element.alias == proxy.alias)) {
+          throw Exception("Api proxy with name '${proxy.alias} is already defined!");
+        }
+        model.proxies.add(proxy);
       } else {
         throw Exception("Unknown node with name: '${node.name.toString()}'");
       }