+++ lib/src/model/entity.dart
@@ -7,9 +7,11 @@ class Entity {
   final List<String> dependencies = [];
   final List<Field> fields = [];
   final List<String> serializers = [];
+  final List<String> apiProxies;
 
   Entity(
       {required this.name,
       required this.copyWith,
-      required this.generateEntity});
+      required this.generateEntity,
+      required this.apiProxies});
 }