+++ lib/src/writer/entity_writer.dart
@@ -1,3 +1,5 @@
+import 'package:entity_serializer/src/writer/api_proxy_writer.dart';
+
 import '../model/entity.dart';
 import '../model/field.dart';
 import 'import_writer.dart';
@@ -9,8 +11,7 @@ class EntityWriter {
 
   void collectExternalImports(ImportWriter collector) {
     if (entity.copyWith) {
-      collector
-          .addImport("package:copy_with_extension/copy_with_extension.dart");
+      collector.addImport("package:copy_with_extension/copy_with_extension.dart");
     }
   }
 
@@ -28,13 +29,12 @@ class EntityWriter {
     String Function(String name) createEntityFilePath,
   ) {
     if (entity.copyWith) {
-      final path =
-          createEntityFilePath(entity.name).replaceFirst(".dart", ".g.dart");
+      final path = createEntityFilePath(entity.name).replaceFirst(".dart", ".g.dart");
       collector.addPart(path);
     }
   }
 
-  void writeBody(StringBuffer buffer) {
+  void writeBody({required StringBuffer buffer, required ApiProxyWriter proxiesWriter}) {
     if (entity.copyWith) {
       buffer.writeln("@CopyWith()");
     }
@@ -44,6 +44,10 @@ class EntityWriter {
     }
     buffer.writeln();
     _writeConstructor(entity, buffer);
+    buffer.writeln();
+    for(final proxyAlias in entity.apiProxies) {
+      proxiesWriter.writeProxyMethods(entity: entity, buffer: buffer, proxyAlias: proxyAlias);
+    }
     buffer.writeln('}\n');
   }
 
@@ -56,8 +60,7 @@ class EntityWriter {
     if (f.isList) {
       buffer.writeln("  ${fStr}List<${f.valueType}>$optStr ${f.name};");
     } else if (f.isMap) {
-      buffer.writeln(
-          "  ${fStr}Map<${f.keyType}, ${f.valueType}>$optStr ${f.name};");
+      buffer.writeln("  ${fStr}Map<${f.keyType}, ${f.valueType}>$optStr ${f.name};");
     } else {
       //plain field
       buffer.writeln("  $fStr${f.type}$optStr ${f.name};");